refactor: edit api files structure, add token cache
